+++ __tests__/device/takeScreenshot.spec.js
@@ -5,6 +5,7 @@ const { NotImplementedError } = require("../../src/errors");
 
 afterEach(() => {
   appiumServer.resetMocks();
+  jest.restoreAllMocks();
 });
 
 it("takes a screenshot and stores on disk at the given 'filePath' location", async () => {
@@ -29,6 +30,18 @@ it("correctly handles screenshot request errors", async () => {
     .rejects.toThrow(new Error("Failed to take screenshot."));
 
   expect(appiumServer.getCalls(screenshotMock)).toHaveLength(1);
+});
+
+it("correct handles file system errors", async () => {
+  const screenshotMock = appiumServer.mockScreenshot({ data: "dGVzdA==" });
+  const writeFileSpy = jest.spyOn(fs, "writeFile")
+    .mockImplementation((path, data, cb) => cb(new Error("File System error.")));
+
+  await expect(device.takeScreenshot({ filePath: "some/path" }))
+    .rejects.toThrow(new Error("File System error."));
+
+  expect(appiumServer.getCalls(screenshotMock)).toHaveLength(1);
+  expect(writeFileSpy).toHaveBeenCalled();
 });
 
 // TODO: Until we figure out what makes the most sense.